Added Roles Context Example
Added the roles context example as a way to implement the basic fetch a page of data from API with a separate context in a specific page.
Refactored the adapter to have page meta data and added a type to the FetchResult to reduce generic results and allow the context to handle the errors and any response data (such as page meta data).