Refactor Router and Session Context for Init
Removed the Private Route component in favor of a simpler initial calculation for session restore process and rendering of routes after complete.
Adjusted the auth API to split user from token and refresh token calls. Let the session context handle the state of restoring a session and keep it out of the API.
Small refactor on entities and mixing to split out the Restful from the paged and restorable entities. Applied the adjustment to roles and users.
Cleaned up login page and added link to help to demonstrated that help is not a protected route.